Refactored common code from sort versions into core headers
