+++ include/algorithms/dbops/sort/sort_indirect_inplace_cluster_leaf.hpp
@@ -33,6 +33,7 @@
 #include <type_traits>
 
 #include "algorithms/dbops/dbops_hints.hpp"
+#include "algorithms/dbops/sort/sort_core_inplace.hpp"
 #include "algorithms/dbops/sort/sort_utils.hpp"
 #include "algorithms/utils/hinting.hpp"
 #include "algorithms/utils/sorthints.hpp"
@@ -66,7 +67,7 @@ namespace tuddbs {
 
     auto operator()(const size_t left, const size_t right) {
       if ((right - left) < (4 * SimdStyle::vector_element_count())) {
-        insertion_sort_fallback<SortOrderT>(m_data, m_idx, left, right);
+        sort_inplace::insertion_sort_fallback<SortOrderT>(m_data, m_idx, left, right);
         detect_cluster(clusters, m_data, m_idx, left, right);
         return;
       }
@@ -78,156 +79,6 @@ namespace tuddbs {
     std::deque<tuddbs::Cluster> getClusters() const { return clusters; }
 
    private:
-    template <typename T, typename U>
-    void detect_cluster(std::deque<Cluster>& clusters, T* data, U* indexes, size_t left, size_t right) {
-      T run_value = data[left];
-      size_t run_length = 1;
-      ++left;
-      if (left > right) {
-        return;
-      }
-      for (; left < right; ++left) {
-        const T current_value = data[left];
-        if (current_value != run_value) {
-          if (run_length > 1) {
-            clusters.emplace_back(left - run_length, run_length);
-          }
-          run_length = 1;
-          run_value = current_value;
-        } else {
-          ++run_length;
-        }
-      }
-      if (run_length > 1) {
-        clusters.emplace_back(left - run_length, run_length);
-      }
-    }
-
-    template <class S, class I>
-    constexpr static size_t idx_arr_len = sizeof(typename I::base_type) / sizeof(typename S::base_type);
-
-    template <class S, class I>
-    constexpr static size_t bits_per_idx_register =
-      (I::vector_element_count() == 64) ? -1ull : (1ull << I::vector_element_count()) - 1;
-
-    template <class S, class I>
-    using idx_arr_t = std::array<typename I::base_type, S::vector_element_count()>;
-
-    template <class SimdStyle, class IndexStyle, typename U = typename IndexStyle::base_type>
-    inline void compress_store_index_array(typename SimdStyle::imask_type full_mask, U* indexes,
-                                           const idx_arr_t<SimdStyle, IndexStyle> idx_tmparr) {
-#pragma unroll(idx_arr_len<SimdStyle, IndexStyle>)
-      for (size_t i = 0; i < idx_arr_len<SimdStyle, IndexStyle>; ++i) {
-        /**
-         * No binary masking necessary, as we guarantee that the IndexStyle::base_type is equally sized or wider
-         * than SimdStyle::base_type. Thus, IndexStyle::imask_type will always be smaller or equally sized than
-         * SimdStyle::base_type --> Upper/Unnecessary bits will be automatically truncated.
-         */
-        const typename IndexStyle::imask_type idx_mask = static_cast<typename IndexStyle::imask_type>(
-          (full_mask >> (i * IndexStyle::vector_element_count())) & bits_per_idx_register<SimdStyle, IndexStyle>);
-        tsl::compress_store<IndexStyle>(idx_mask, indexes,
-                                        tsl::loadu<IndexStyle>(&idx_tmparr[i * IndexStyle::vector_element_count()]));
-        /**
-         * We always 'just' advance the indexes pointer by the amount of written elements, both for left and right
-         * side. That is, for left we always start at the 'first' write position. For right, we position the indexes
-         * pointer at the leftmost position, since we know nb_high from the full_mask.
-         */
-        indexes += tsl::mask_population_count<IndexStyle>(idx_mask);
-      }
-    }
-
-    template <class SimdStyle, SORT_TYPE type, tuddbs::TSL_SORT_ORDER order>
-    __attribute__((always_inline)) inline auto compare(const typename SimdStyle::register_type val_reg,
-                                                       const typename SimdStyle::register_type pivot_reg) ->
-      typename SimdStyle::imask_type {
-      if constexpr (type == SORT_TYPE::SORT_EQ) {
-        return tsl::to_integral<SimdStyle>(tsl::equal<SimdStyle>(val_reg, pivot_reg));
-      } else {
-        if constexpr (order == TSL_SORT_ORDER::ASC) {
-          return tsl::to_integral<SimdStyle>(tsl::less_than<SimdStyle>(val_reg, pivot_reg));
-        } else {
-          return tsl::to_integral<SimdStyle>(tsl::greater_than<SimdStyle>(val_reg, pivot_reg));
-        }
-      }
-    }
-
-    template <class SimdStyle, SORT_TYPE type, TSL_SORT_ORDER order>
-    __attribute__((always_inline)) inline auto compare(const typename SimdStyle::register_type val_reg,
-                                                       const typename SimdStyle::register_type pivot_reg,
-                                                       const typename SimdStyle::imask_type valid) ->
-      typename SimdStyle::imask_type {
-      return compare<SimdStyle, type, order>(val_reg, pivot_reg) & valid;
-    }
-
-    template <class SimdStyle, class IndexStyle, SORT_TYPE type, TSL_SORT_ORDER order,
-              typename T = typename SimdStyle::base_type, typename U = typename IndexStyle::base_type>
-    inline void do_tsl_sort(T* data, U* indexes, const typename SimdStyle::register_type pivot_reg,
-                            const typename SimdStyle::register_type val_reg,
-                            const idx_arr_t<SimdStyle, IndexStyle> idx_tmparr, ssize_t& l_w, ssize_t& r_w) {
-      using mask_t = typename SimdStyle::imask_type;
-      const mask_t mask_lt = compare<SimdStyle, type, order>(val_reg, pivot_reg);
-      const size_t nb_low = tsl::mask_population_count<SimdStyle>(mask_lt);
-      const size_t nb_high = SimdStyle::vector_element_count() - nb_low;
-
-      tsl::compress_store<SimdStyle>(mask_lt, &data[l_w], val_reg);
-      compress_store_index_array<SimdStyle, IndexStyle>(mask_lt, &indexes[l_w], idx_tmparr);
-      l_w += nb_low;
-      r_w -= nb_high;
-      tsl::compress_store<SimdStyle>(static_cast<mask_t>(~mask_lt), &data[r_w], val_reg);
-      compress_store_index_array<SimdStyle, IndexStyle>(static_cast<mask_t>(~mask_lt), &indexes[r_w], idx_tmparr);
-    }
-
-    template <class SimdStyle, class IndexStyle, SORT_TYPE type, TSL_SORT_ORDER order,
-              typename T = typename SimdStyle::base_type, typename U = typename IndexStyle::base_type>
-    inline void do_tsl_sort_masked(T* data, U* indexes, const typename SimdStyle::register_type pivot_reg,
-                                   const typename SimdStyle::register_type val_reg,
-                                   const idx_arr_t<SimdStyle, IndexStyle> idx_tmparr, ssize_t& l_w, ssize_t& r_w,
-                                   const typename SimdStyle::imask_type valid) {
-      using mask_t = typename SimdStyle::imask_type;
-
-      const mask_t mask_lt = compare<SimdStyle, type, order>(val_reg, pivot_reg, valid);
-      const mask_t mask_gt = static_cast<mask_t>(~mask_lt) & valid;
-
-      const size_t nb_low = tsl::mask_population_count<SimdStyle>(mask_lt);
-      const size_t nb_high = tsl::mask_population_count<SimdStyle>(mask_gt);
-
-      tsl::compress_store<SimdStyle>(mask_lt, &data[l_w], val_reg);
-      compress_store_index_array<SimdStyle, IndexStyle>(mask_lt, &indexes[l_w], idx_tmparr);
-      l_w += nb_low;
-      r_w -= nb_high;
-      tsl::compress_store<SimdStyle>(mask_gt, &data[r_w], val_reg);
-      compress_store_index_array<SimdStyle, IndexStyle>(mask_gt, &indexes[r_w], idx_tmparr);
-    }
-
-    template <TSL_SORT_ORDER order, typename T, typename U>
-    void insertion_sort_fallback(T* data, U* indexes, const ssize_t left_boundary, const ssize_t right_boundary) {
-      const size_t idxs_count = (right_boundary - left_boundary);
-      ssize_t i, j;
-      T val;
-      U idx;
-      for (i = left_boundary + 1; i < right_boundary; i++) {
-        val = data[i];
-        idx = indexes[i];
-        j = i - 1;
-
-        if constexpr (order == TSL_SORT_ORDER::ASC) {
-          while (j >= 0 && data[j] > val) {
-            data[j + 1] = data[j];
-            indexes[j + 1] = indexes[j];
-            j = j - 1;
-          }
-        } else {
-          while (j >= 0 && data[j] < val) {
-            data[j + 1] = data[j];
-            indexes[j + 1] = indexes[j];
-            j = j - 1;
-          }
-        }
-        data[j + 1] = val;
-        indexes[j + 1] = idx;
-      }
-    };
-
     template <class SimdStyle, class IndexStyle, typename T = typename SimdStyle::base_type,
               typename U = typename IndexStyle::base_type>
     void partition(std::deque<Cluster>& cluster, T* data, U* indexes, ssize_t left, ssize_t right, T pivot) {
@@ -279,30 +130,30 @@ namespace tuddbs {
       idx_arr_t<SimdStyle, IndexStyle> idxs;
 
       if (left == right) {
-        do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_l, idx_l,
-                                                                           left_w, right_w);
-        do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_l_adv,
-                                                                           idx_l_adv, left_w, right_w);
-        do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_r, idx_r,
-                                                                           left_w, right_w);
-        do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_r_adv,
-                                                                           idx_r_adv, left_w, right_w);
+        sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+          data, indexes, pivot_vec, vals_l, idx_l, left_w, right_w);
+        sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+          data, indexes, pivot_vec, vals_l_adv, idx_l_adv, left_w, right_w);
+        sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+          data, indexes, pivot_vec, vals_r, idx_r, left_w, right_w);
+        sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+          data, indexes, pivot_vec, vals_r_adv, idx_r_adv, left_w, right_w);
       } else if (left + SimdStyle::vector_element_count() > right) {
         const typename SimdStyle::imask_type valid_mask = ((1ull << (right - left)) - 1) & (-1ull >> 1);
         const data_reg_t remainder_val_vec = tsl::loadu<SimdStyle>(&data[left]);
         const idx_arr_t<SimdStyle, IndexStyle> remainder_idx_vec = load_idx_arr(&indexes[left]);
 
-        do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_l, idx_l,
-                                                                           left_w, right_w);
-        do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_l_adv,
-                                                                           idx_l_adv, left_w, right_w);
-        do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_r, idx_r,
-                                                                           left_w, right_w);
-        do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_r_adv,
-                                                                           idx_r_adv, left_w, right_w);
+        sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+          data, indexes, pivot_vec, vals_l, idx_l, left_w, right_w);
+        sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+          data, indexes, pivot_vec, vals_l_adv, idx_l_adv, left_w, right_w);
+        sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+          data, indexes, pivot_vec, vals_r, idx_r, left_w, right_w);
+        sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+          data, indexes, pivot_vec, vals_r_adv, idx_r_adv, left_w, right_w);
 
         /* Remainder */
-        do_tsl_sort_masked<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+        sort_inplace::do_tsl_sort_masked<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
           data, indexes, pivot_vec, remainder_val_vec, remainder_idx_vec, left_w, right_w, valid_mask);
       } else {
         while (left + SimdStyle::vector_element_count() <= right) {
@@ -331,8 +182,8 @@ namespace tuddbs {
             idx_r_adv = load_idx_arr(&indexes[right]);
           }
 
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals, idxs,
-                                                                             left_w, right_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec,
+                                                                                           vals, idxs, left_w, right_w);
         }
 
         /* -- Cleanup Phase -- */
@@ -341,27 +192,27 @@ namespace tuddbs {
           const data_reg_t remainder_val_vec = tsl::loadu<SimdStyle>(&data[left]);
           const idx_arr_t<SimdStyle, IndexStyle> remainder_idx_vec = load_idx_arr(&indexes[left]);
 
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_l, idx_l,
-                                                                             left_w, right_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_l_adv,
-                                                                             idx_l_adv, left_w, right_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_r, idx_r,
-                                                                             left_w, right_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_r_adv,
-                                                                             idx_r_adv, left_w, right_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+            data, indexes, pivot_vec, vals_l, idx_l, left_w, right_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+            data, indexes, pivot_vec, vals_l_adv, idx_l_adv, left_w, right_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+            data, indexes, pivot_vec, vals_r, idx_r, left_w, right_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+            data, indexes, pivot_vec, vals_r_adv, idx_r_adv, left_w, right_w);
 
           /* Remainder */
-          do_tsl_sort_masked<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+          sort_inplace::do_tsl_sort_masked<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
             data, indexes, pivot_vec, remainder_val_vec, remainder_idx_vec, left_w, right_w, valid_mask);
         } else {
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_l, idx_l,
-                                                                             left_w, right_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_l_adv,
-                                                                             idx_l_adv, left_w, right_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_r, idx_r,
-                                                                             left_w, right_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(data, indexes, pivot_vec, vals_r_adv,
-                                                                             idx_r_adv, left_w, right_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+            data, indexes, pivot_vec, vals_l, idx_l, left_w, right_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+            data, indexes, pivot_vec, vals_l_adv, idx_l_adv, left_w, right_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+            data, indexes, pivot_vec, vals_r, idx_r, left_w, right_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_LT, SortOrderT>(
+            data, indexes, pivot_vec, vals_r_adv, idx_r_adv, left_w, right_w);
         }
       }
 
@@ -373,7 +224,7 @@ namespace tuddbs {
       ssize_t pivot_r_w = right_start;
       if ((pivot_r - pivot_l) < 4 * SimdStyle::vector_element_count()) {
         /* Data too small to fit in 4 registers */
-        insertion_sort_fallback<SortOrderT>(data, indexes, pivot_l, pivot_r);
+        sort_inplace::insertion_sort_fallback<SortOrderT>(data, indexes, pivot_l, pivot_r);
 
         for (size_t i = pivot_l; i < right_start; ++i) {
           if (data[i] > pivot) {
@@ -399,30 +250,30 @@ namespace tuddbs {
         idx_r_adv = load_idx_arr(&indexes[pivot_r]);
 
         if (pivot_l == pivot_r) {
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_l, idx_l,
-                                                                             pivot_l_w, pivot_r_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_l_adv,
-                                                                             idx_l_adv, pivot_l_w, pivot_r_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_r, idx_r,
-                                                                             pivot_l_w, pivot_r_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_r_adv,
-                                                                             idx_r_adv, pivot_l_w, pivot_r_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+            data, indexes, pivot_vec, vals_l, idx_l, pivot_l_w, pivot_r_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+            data, indexes, pivot_vec, vals_l_adv, idx_l_adv, pivot_l_w, pivot_r_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+            data, indexes, pivot_vec, vals_r, idx_r, pivot_l_w, pivot_r_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+            data, indexes, pivot_vec, vals_r_adv, idx_r_adv, pivot_l_w, pivot_r_w);
         } else if (pivot_l + SimdStyle::vector_element_count() > pivot_r) {
           const typename SimdStyle::imask_type valid_mask = ((1ull << (pivot_r - pivot_l)) - 1) & (-1ull >> 1);
           const data_reg_t remainder_val_vec = tsl::loadu<SimdStyle>(&data[pivot_l]);
           const idx_arr_t<SimdStyle, IndexStyle> remainder_idx_vec = load_idx_arr(&indexes[pivot_l]);
 
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_l, idx_l,
-                                                                             pivot_l_w, pivot_r_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_l_adv,
-                                                                             idx_l_adv, pivot_l_w, pivot_r_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_r, idx_r,
-                                                                             pivot_l_w, pivot_r_w);
-          do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_r_adv,
-                                                                             idx_r_adv, pivot_l_w, pivot_r_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+            data, indexes, pivot_vec, vals_l, idx_l, pivot_l_w, pivot_r_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+            data, indexes, pivot_vec, vals_l_adv, idx_l_adv, pivot_l_w, pivot_r_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+            data, indexes, pivot_vec, vals_r, idx_r, pivot_l_w, pivot_r_w);
+          sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+            data, indexes, pivot_vec, vals_r_adv, idx_r_adv, pivot_l_w, pivot_r_w);
 
           /* Remainder */
-          do_tsl_sort_masked<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+          sort_inplace::do_tsl_sort_masked<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
             data, indexes, pivot_vec, remainder_val_vec, remainder_idx_vec, pivot_l_w, pivot_r_w, valid_mask);
         } else {
           while (pivot_l + SimdStyle::vector_element_count() <= pivot_r) {
@@ -451,8 +302,8 @@ namespace tuddbs {
               idx_r_adv = load_idx_arr(&indexes[pivot_r]);
             }
 
-            do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals, idxs,
-                                                                               pivot_l_w, pivot_r_w);
+            sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+              data, indexes, pivot_vec, vals, idxs, pivot_l_w, pivot_r_w);
           }
 
           /* -- Cleanup Phase -- */
@@ -461,34 +312,34 @@ namespace tuddbs {
             const data_reg_t remainder_val_vec = tsl::loadu<SimdStyle>(&data[pivot_l]);
             const idx_arr_t<SimdStyle, IndexStyle> remainder_idx_vec = load_idx_arr(&indexes[pivot_l]);
 
-            do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_l, idx_l,
-                                                                               pivot_l_w, pivot_r_w);
-            do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_l_adv,
-                                                                               idx_l_adv, pivot_l_w, pivot_r_w);
-            do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_r, idx_r,
-                                                                               pivot_l_w, pivot_r_w);
-            do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_r_adv,
-                                                                               idx_r_adv, pivot_l_w, pivot_r_w);
+            sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+              data, indexes, pivot_vec, vals_l, idx_l, pivot_l_w, pivot_r_w);
+            sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+              data, indexes, pivot_vec, vals_l_adv, idx_l_adv, pivot_l_w, pivot_r_w);
+            sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+              data, indexes, pivot_vec, vals_r, idx_r, pivot_l_w, pivot_r_w);
+            sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+              data, indexes, pivot_vec, vals_r_adv, idx_r_adv, pivot_l_w, pivot_r_w);
 
             /* Remainder */
-            do_tsl_sort_masked<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+            sort_inplace::do_tsl_sort_masked<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
               data, indexes, pivot_vec, remainder_val_vec, remainder_idx_vec, pivot_l_w, pivot_r_w, valid_mask);
           } else {
-            do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_l, idx_l,
-                                                                               pivot_l_w, pivot_r_w);
-            do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_l_adv,
-                                                                               idx_l_adv, pivot_l_w, pivot_r_w);
-            do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_r, idx_r,
-                                                                               pivot_l_w, pivot_r_w);
-            do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(data, indexes, pivot_vec, vals_r_adv,
-                                                                               idx_r_adv, pivot_l_w, pivot_r_w);
+            sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+              data, indexes, pivot_vec, vals_l, idx_l, pivot_l_w, pivot_r_w);
+            sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+              data, indexes, pivot_vec, vals_l_adv, idx_l_adv, pivot_l_w, pivot_r_w);
+            sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+              data, indexes, pivot_vec, vals_r, idx_r, pivot_l_w, pivot_r_w);
+            sort_inplace::do_tsl_sort<SimdStyle, IndexStyle, SORT_TYPE::SORT_EQ, SortOrderT>(
+              data, indexes, pivot_vec, vals_r_adv, idx_r_adv, pivot_l_w, pivot_r_w);
           }
         }
       }
 
       /* -- Left Side -- */
       if ((left_w - left_start) < (4 * SimdStyle::vector_element_count())) {
-        insertion_sort_fallback<SortOrderT>(data, indexes, left_start, left_w);
+        sort_inplace::insertion_sort_fallback<SortOrderT>(data, indexes, left_start, left_w);
         detect_cluster(cluster, data, indexes, left_start, left_w);
       } else {
         const auto pivot_ls = get_pivot(data, left_start, left_w);
@@ -497,7 +348,7 @@ namespace tuddbs {
 
       /* -- Right Side -- */
       if ((right_start - pivot_r_w) < (4 * SimdStyle::vector_element_count())) {
-        insertion_sort_fallback<SortOrderT>(data, indexes, right_w, right_start);
+        sort_inplace::insertion_sort_fallback<SortOrderT>(data, indexes, right_w, right_start);
         detect_cluster(cluster, data, indexes, left_w, right_start);
       } else {
         const auto pivot_rs = get_pivot(data, pivot_r_w, right_start);
